Delete ansible plan when delete plan cr
* Delete ansible plan when delete plan cr.
* Add the retry mechanism when executing ansible plan.
