+++ pkg/utils/ansible.go
@@ -47,6 +47,10 @@ kube-node
 {{range .KubeLog}}
 {{.}}
 {{end}}
+[nvidia-accelerator]
+{{range .NvidiaAccelerator}}
+{{.}}
+{{end}}
 {{range $key, $value := .OtherGroup}}
 [{{$key}}]
 {{range $value}}
